Create canvas for Custom plot creation
